DATAES-290 Make joda-time optional dependency.
Make the joda-time optional dependency. Also check if joda-time is in the classpath before registering
the joda-time converters.

This is the first step towards making joda-time as optional dependency. However it is also being used in FacetedPageImpl.java#L161 which
is deprecated and needs to be removed before making the joda-time truly optional.