Support collection parameters in @Query methods
This adds a test and fix so that collection parameters are supported and quoted correctly, at least for `Collection<String>`.

Previously the `StringQueryUtil` class would just call `.toString()` on the parameter and escape the `"` characters in the result, which would cause a syntax error in the query string.

This changes so that the class checks if the parameter is a `Collection` and if so starts a Json array and adds elements to it by calling `convert()` on each element.